Fixed places that used shadersManager
Former-commit-id: c2acff86c39a2fe033ffe2e25c30cbe7691a3c5f
Former-commit-id: 28f855a5d8179db73f2daef54da95ab58da22893

+++ src/uiRendering/controlRenderer.h
@@ -8,6 +8,7 @@
 #include <rendering\shader.h>
 #include <rendering\buffer.h>
 #include <rendering\vertexBuffer.h>
+#include <rendering\program.h>
 
 #include <ui\control.h>
 
@@ -20,7 +21,7 @@ namespace phi
     private:
         struct controlInstance
         {
-            uint bufferOffset;
+            size_t bufferOffset;
             mat4 modelMatrix;
             controlRenderData renderData;
         };
